cwhisper: fix old interval trimming bug
getBufferInfo is always returning min as 0 and this effectively renders the old internval
check (minInterval != 0 && bpBaseInterval < minInterval) in archiveUpdateManyCompressed
useless. The impact is old timestamps could be saved, this could caused incorrect read
result. But in theory this should only happen to metrics that is receving out of order
data points and the gap is larger than the buffer (if a metric is being generated by a
single data source, then it should to be fine).